ConfigProvider support `autoInsertSpaceInButton` (#14230)
close #14229
